requirements.txt
Flask==2.3.1
gtfs-realtime-bindings==1.0.0
protobuf==4.23.2
requests>=2.30.0
simplejson==3.17.6
swagger-ui-py==22.7.13
wget==3.2
warsaw-data-api==0.3.5
zipp==1.0.0
gunicorn==21.2.0